feat: Add directoryListingPatterns to checkNonHTMLContentAccessibility module
This commit adds the `directoryListingPatterns` variable to the `checkNonHTMLContentAccessibility` module in the `checkNonHtmlAccccesability.js` file. The variable is exported along with the `checkNonHTMLContentAccessibility` function. This change allows the function to access the patterns for checking directory listings in non-HTML content.

Refactor InsecureObjectRefGuard.controoler.js to import directoryListingPatterns

This commit refactors the `InsecureObjectRefGuard.controoler.js` file to import the `directoryListingPatterns` variable from the `checkNonHtmlAccccesability` module. The variable is used to check if the page content includes any of the patterns for directory listings. The appropriate response is sent based on the result.

Update scanClientData.js to fix formatting and remove unnecessary code

This commit updates the `scanClientData.js` file to fix formatting issues and remove unnecessary code. The changes include fixing indentation, removing extra spaces, and improving code readability.

Merge pull request #20 from shivam-handsintechnology/shivam

This commit merges the changes from the `shivam-handsintechnology/shivam` branch into the main branch.

backup

This commit is a backup.

feat: Add endpoint for managing web domains

This commit adds a new endpoint '/webdomain' to the security routes for managing web domains. The endpoint supports GET and POST methods for retrieving and adding web domains respectively. It also includes the necessary controller functions and validation middleware.

refactor: Remove unused code for sensitive data handling

This commit removes unused code related to sensitive data handling. The code is no longer needed and can be safely removed.

feat: Add accessibility testing for non-HTML content

This commit adds accessibility testing for non-HTML content. It includes a new function `scanNonHTMLContentAccessibility` in the `scanClientData.js` file. The function checks for the presence of forms and elements with a class containing "modal" on a page.

chore: Add session middleware for user authentication and data persistence

This commit adds session middleware for user authentication and data persistence. It includes the necessary code in the `AuthSessionGuardian.route.js` file.

Refactor protect.js to increase interval for sending data to API

This commit refactors the `protect.js` file to increase the interval for sending data to the API. The interval is adjusted to improve performance and reduce unnecessary API calls.

Refactor protect.js to handle localStorage changes and send data to API

This commit refactors the `protect.js` file to handle localStorage changes and send data to the API. The necessary code is added to detect changes in localStorage and send the updated data to the API.